pyproject.toml
[project]
name = "outline-generation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.7"
dependencies = [
    "elevenlabs>=2.34.0",
    "jinja2>=3.1.6",
    "openai>=2.17.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "streamlit>=1.42.0",
]
